+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
  - Add Math Controller (Math in menu) to perform math on Input data
+ - Add first Math controller type: Average (averages Inputs) ([#328](https://github.com/kizniche/mycodo/issues/328))
  - Add fswebcam as a camera library for acquiring images from USB cameras
  - Complete Spanish translation
  - Update korean translations